More tests using the real database service factory
