+++ apiserver/facades/client/application/application_test.go
@@ -73,8 +73,11 @@ func (s *applicationSuite) makeAPI(c *gc.C) *application.APIBase {
 	model, err := st.Model()
 	c.Assert(err, jc.ErrorIsNil)
 	blockChecker := common.NewBlockChecker(st)
+
+	serviceFactory := s.ServiceFactory(jujutesting.DefaultModelUUID)
+
 	env, err := stateenvirons.GetNewEnvironFunc(
-		environs.New)(s.ControllerModel(c), s.ControllerServiceFactory.Cloud(), s.ControllerServiceFactory.Credential())
+		environs.New)(s.ControllerModel(c), serviceFactory.Cloud(), serviceFactory.Credential())
 	c.Assert(err, jc.ErrorIsNil)
 	registry := stateenvirons.NewStorageProviderRegistry(env)
 	pm := poolmanager.New(state.NewStateSettings(st), registry)
@@ -87,8 +90,8 @@ func (s *applicationSuite) makeAPI(c *gc.C) *application.APIBase {
 		nil,
 		blockChecker,
 		application.GetModel(model),
-		s.ControllerServiceFactory.Cloud(),
-		s.ControllerServiceFactory.Credential(),
+		serviceFactory.Cloud(),
+		serviceFactory.Credential(),
 		nil, // leadership not used in these tests.
 		application.CharmToStateCharm,
 		application.DeployApplication,
